+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add i18n to profile description.
 - Add i18n to accommodation description.
 - Show version and commit hash on homepage.
+- Setup e2e tests with Playwright.
 
 ### Changed